+++ cypress/e2e/result_panel.cy.js
@@ -1,0 +1,30 @@
+import "cypress-real-events";
+describe("Test - Import JSON", () => {
+    it("Pick one JSON file and check trasnfert status", () => {
+        let upName = "Test E2E Usage Pattern";
+
+        cy.visit("/");
+        cy.get('#btn-start-modeling-my-service').click();
+        cy.get('#model-canva').should('be.visible');
+        cy.get('button[hx-get="/model_builder/open-import-json-panel/"]').click();
+        let fileTest = 'cypress/fixtures/efootprint-model-system-data.json'
+        cy.get('input[type="file"]').selectFile(fileTest);
+        cy.get('input[type="file"]').then(($input) => {
+          expect($input[0].files.length).to.equal(1);
+          expect($input[0].files[0].name).to.equal('efootprint-model-system-data.json');
+        });
+        cy.get('button[type="submit"]').click();
+        cy.wait(1000);
+        cy.get('button[id^="button-id-"][id$="'+upName.replaceAll(' ', '-')+'"]').should('exist').should('be.visible');
+
+        cy.get('#btn-open-panel-result')
+        .realTouch('start', { x: 100, y: 300 })
+        .realTouch('move', { x: 100, y: 200 })
+        .realTouch('end', { x: 100, y: 200 });
+        cy.wait(2000);
+        cy.get('#inner-panel-result').should('be.visible').find('div[onclick="hidePanelResult()"]')
+        .realTouch('start', { x: 100, y: 300 })
+        .realTouch('move', { x: 100, y: 400 })
+        .realTouch('end', { x: 100, y: 400 });
+    });
+});