+++ theme/static/scripts/model_builder_main.js
@@ -18,6 +18,8 @@ function initSortableObjectCards() {
 function initModelBuilderMain() {
     initLeaderLines();
     initSortableObjectCards();
+    initSidePanel();
+    initHammer();
 }
 
 function reverseIconAccordion(objectId){